Add email disclaimer to register page
Emails are not public in the app, and Google requires a login mehtod
where the user opts out of their email being public. This way,
the user is aware of their privacy in our application.

With the profile/testing group update, we will need to make the user
accept that once they opt in to that feature, their emails will be
public