Change KeyBoardAvoiding view to 'position' for iOS
The behavior sometimes caused issues when switching to BO3 in the
form (changing the height), and some users have reported freezes
when doing so. The freezes were not recreatable, but we can assume
this buggy behavior was related.

Position is less nice looking, but it works and has not caused
any glitchy UIs in testing